+++ src/views/table/BulkActionBar.ts
@@ -3,7 +3,7 @@ import type { Task, TaskStatus, TaskPriority } from '../../types'
 import { flattenTasks, collectAllAssignees, collectAllTags } from '../../store'
 import { displayName, formatBadgeText } from '../../utils'
 import { today } from '../../dates'
-import { promptText } from '../../ui/ModalFactory'
+import { openPersonPicker, promptText } from '../../ui/ModalFactory'
 import { TaskPickerModal } from '../../modals/PickerModals'
 import type { TableContext } from './TableRenderer'
 import { updateSelectAllCheckbox } from './TableRow'
@@ -88,9 +88,10 @@ function updateBarContent(bar: HTMLElement, ctx: TableContext, onAction: (a: Bul
     }
     menu.addSeparator()
     menu.addItem((item) =>
-      item.setTitle('+ new assignee...').onClick(async () => {
-        const name = await promptText(ctx.plugin.app, 'Enter assignee name:', 'Name')
-        if (name) onAction({ type: 'set-assignee', assignee: name })
+      item.setTitle('Someone else...').onClick(() => {
+        openPersonPicker(ctx.plugin, allMembers, ctx.scope.primary?.filePath ?? '', (value) => {
+          onAction({ type: 'set-assignee', assignee: value })
+        })
       })
     )
     menu.addSeparator()